Add error codes Switch to using Result in io functions Remove outdated ZFS scheme, as the ZFS app has new development

+++ filesystem/apps/zfs/zio.rs
@@ -15,14 +15,14 @@ impl Reader {
     pub fn read(&mut self, start: usize, length: usize) -> Vec<u8> {
         let mut ret: Vec<u8> = vec![0; length*512];
 
-        self.disk.seek(SeekFrom::Start(start * 512));
+        self.disk.seek(SeekFrom::Start(start as u64 * 512));
         self.disk.read(&mut ret);
 
         return ret;
     }
 
     pub fn write(&mut self, block: usize, data: &[u8; 512]) {
-        self.disk.seek(SeekFrom::Start(block * 512));
+        self.disk.seek(SeekFrom::Start(block as u64 * 512));
         self.disk.write(data);
     }
